tutorial/requirement.txt
django==1.7.2
djangorestframework
mysql-python




